+++ src/features/exhibition/exhibitoinList/EXListBody.jsx
@@ -4,9 +4,10 @@ import tickets from "../../../assets/imgs/exhibition/tickets.png";
 import artist from "../../../assets/imgs/exhibition/artist.png";
 import sparkle_gray from "../../../assets/imgs/exhibition/sparkle_gray.png";
 import { useNavigate } from "react-router-dom";
+import { ExSatus } from "../EXstatusTitle";
 export const EXListBody = ({ isLoading, isError, merged }) => {
   const navigator = useNavigate();
-  console.log(merged.filter(el => el.endDate === null));
+  console.log(merged.filter((el) => el.endDate === null));
   return (
     <>
       {isLoading || isError ? (
@@ -23,13 +24,7 @@ export const EXListBody = ({ isLoading, isError, merged }) => {
                 {item.address?.split(" ").slice(0, 2).join(" ")}
               </ExhibitionDate>
               <ExSate>
-                {item.exhibitionStatus === "전시 진행"
-                  ? "Now On View"
-                  : item.exhibitionStatus === "전시 예정"
-                  ? "Coming Soon"
-                  : item.exhibitionStatus === "전시 종료"
-                  ? "Exhibition is over"
-                  : ""}
+                <ExSatus info={item} />
               </ExSate>
               <ExhibitionTitleWrap>
                 <Excategoryname>{item.categoryCodeName[0]}</Excategoryname>
